By default append .config to user's module names in cfg-create
The idea is that in the bpkg amalgamation we only want the configuration,
without any rules, etc.

This can be overridden by adding the period at the end of the module name.
While at it, recognize the leading question mark in the module name and treat
it as the optional module load.