Use curl instead of wget as default fetch program
We used to prefer wget 1.16 because it has --show-progress which results in
nicer progress. But experience shows that wget is quite unreliable plus with
bdep always using curl, it would be strange to use both curl and wget (and
expecting the user to setup proxy, authentication, etc., for both).

Also add internal --curl* options to be used by bdep.