Fix collect_dependents() to not add existing dependents as being built ones
Also relax the try_replace_dependency() function to accept not only the being
build packages (build action) but the existing packages (adjust action) as
well. Also make it to return unsatisfied dependents of both these action
types. This, in particular, allows try_replace_dependent() to consider
dependents of both the build and adjust types for replacement, which increases
the chances for unsatisfied dependency constraints to be resolved.